# api.yaml
openapi: 3.0.3
info:
  title: Skydiver
  description: API
  version: LATEST
...
paths:
...
  /product/info:
    get:
      description: Получение информации о продуке по id
      operationId: productInfo
      tags:
        - product
      parameters:
        - name: productId
          in: query
          description: product id
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Успешный поиск продукта
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
  /product/list:
    get:
      description: Ручка с фильтрами для списка продуктов
      operationId: productList
      tags:
        - product
      parameters:
        - name: priceFrom
          in: query
          description: мин цена
          schema:
            type: integer
        - name: priceTo
          in: query
          description: max цена
          schema:
            type: integer
        - name: searchString
          in: query
          description: поисковая строка
          schema:
            type: string
        - name: category
          in: query
          required: true
          schema:
            type: integer
      responses:
        200:
          description: список продуктов
          content:
             application/json:
               schema:
                 $ref: '#/components/schemas/ProductSearchResponse'
  /product/search:
    get:
      description: Поиск продукта по имени
      operationId: productSearch
      tags:
        - product
      parameters:
        - name: searchString
          in: query
          description: поисковая строка
          required: true
          schema:
            type: string
      responses:
        200:
          description: Успешный поиск продукта
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductSearchResponse'
...
components:
  schemas:
...
    ProductSearchResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: '#/components/schemas/Product'
    Product:
      type: object
      properties:
        id:
          description: id продукта
          type: integer
          nullable: false
        name:
          description: имя продукта
          type: string
          nullable: false
        categoryId:
          description: id категории продукта
          type: integer
          nullable: false
        price:
          description: цена
          type: integer
        productImage:
          description: Изображение продукта
          type: string
        productDescription:
          description: Описание продукта
          type: string
...